+++ src/components/App.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactGA from 'react-ga'
 import { createHistory, LocationProvider, Router } from '@reach/router'
-import Notes from './notes/Notes'
-import SignIn from './auth/SignIn'
-import SignUp from './auth/SignUp'
+import Notes from './Notes'
+import SignIn from './SignIn'
+import SignUp from './SignUp'
 import './styles/index.css'
 
 // @ts-ignore
